requirements.txt
pandas
psycopg2
cx_Oracle
python-decouple